requirements.txt
contourpy==1.3.1
cupy-cuda12x==13.4.0
cycler==0.12.1
fastrlock==0.8.3
fonttools==4.56.0
h5py==3.13.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scipy==1.15.2
six==1.17.0
